statically located connection menu for pages
