+++ app/javascript/src/component_connection_menu.js
@@ -1,0 +1,49 @@
+const utilities = require('./utilities');
+const mergeObjects = utilities.mergeObjects;
+const ActivatedMenu = require('./component_activated_menu');
+
+class ConnectionMenu extends ActivatedMenu {
+  constructor($node, config) {
+    super($node, mergeObjects({
+      activator_classname: $node.data("activator-classname"),
+      container_id: $node.data("activated-menu-container-id"),
+      activator_text: $node.data("activator-text")
+    }, config));
+
+    this.container.$node.addClass("ConnectionMenu");
+
+    // Register event handler for selection of menu item.
+    $node.on("menuselect", (event, ui) => {
+      this.selection(event, ui.item);
+    });
+
+    this.activator.$node.addClass("ConnectionMenuActivator");
+    this.container.$node.addClass("ConnectionMenu");
+    this.uuid = $node.data("uuid");
+    this.title = $node.data("title");
+  }
+
+  set addPageAfter(uuid) {
+    this._uuid = uuid;
+  }
+
+  get addPageAfter() {
+    return this._uuid;
+  }
+
+  selection(event, item) {
+    var dialog = this._config.view.pageAdditionDialog;
+    var $form = dialog.$form;
+
+    // Set the 'add_page_here' value to mark point of new page inclusion.
+    // Should be a uuid of previous page or blank if at end of form.
+    utilities.updateHiddenInputOnForm($form, "page[add_page_after]", this.addPageAfter);
+
+    // Then add any required values.
+    utilities.updateHiddenInputOnForm($form, "page[page_type]", item.data("page-type"));
+    utilities.updateHiddenInputOnForm($form, "page[component_type]", item.data("component-type"));
+
+    this._config.view.pageAdditionDialog.open();
+  }
+}
+module.exports = ConnectionMenu
