+++ app/javascript/src/component_flow_item_menu.js
@@ -1,0 +1,121 @@
+
+
+const utilities = require('./utilities');
+const mergeObjects = utilities.mergeObjects;
+const ActivatedMenu = require('./component_activated_menu');
+
+class FlowItemMenu extends ActivatedMenu {
+  constructor($node, config) {
+    super($node, mergeObjects({
+      activator_classname: $node.data("activator-classname"),
+      container_id: $node.data("activated-menu-container-id"),
+      activator_text: $node.data("activator-text")
+    }, config));
+
+    $node.on("menuselect", (event, ui) => {
+      this.selection(event, ui.item);
+    });
+
+    this.activator.$node.addClass("FlowItemMenuActivator");
+    this.container.$node.addClass("FlowItemMenu");
+    this.uuid = $node.data("uuid");
+    this.title = $node.data("title");
+  }
+
+   // Handle item selections on the form step context menu elements.
+  selection(event, item) {
+    var action = item.data("action");
+
+    event.preventDefault();
+    switch(action) {
+      case "preview":
+           this.previewPage(item);
+           break;
+
+      case "add":
+           this.addPage(item);
+           break;
+
+      case "destination":
+           this.changeDestination(item);
+           break;
+
+      case "delete":
+           this.deleteItem(item);
+           break;
+
+      case "delete-api":
+           this.deleteItemApi(item);
+           break;
+
+      default: this.link(item);
+    }
+  }
+
+  link(element) {
+    var $link = element.find("> a");
+    location.href = $link.attr("href");
+  }
+
+  previewPage(element) {
+    var $link = element.find("> a");
+    window.open($link.attr("href"));
+  }
+
+  // Open the views Page Addition Menu
+  addPage(element) {
+    var menu = this._config.view.pageAdditionMenu;
+    menu.addPageAfter = this.uuid;
+    menu.open({
+      my: "left top",
+      at: "right top",
+      of: element
+    });
+  }
+
+  // Open an API request dialog to change destination
+  changeDestination(element) {
+    var view = this._config.view;
+    var $link = element.find("> a");
+    new DialogApiRequest($link.attr("href"), {
+      activator: $link,
+      buttons: [{
+        text: view.text.dialogs.button_change_destination,
+        click: function(dialog) {
+          dialog.$node.find("form").submit();
+        }
+      }, {
+        text: view.text.dialogs.button_cancel
+      }]
+    });
+  }
+
+  // Use standard delete modal to remove
+  deleteItem(element) {
+    var view = this._config.view;
+    var $link = element.find("> a");
+    view.dialogConfirmationDelete.open({
+      heading: view.text.dialogs.heading_delete.replace(/%{label}/, this.title),
+      ok: view.text.dialogs.button_delete_page
+      }, function() {
+        post($link.attr("href"), { _method: "delete" });
+    });
+  }
+
+  deleteItemApi(element) {
+    var $link = element.find("> a");
+    new DialogApiRequest($link.attr("href"), {
+      activator: $link,
+      closeOnClickSelector: ".govuk-button",
+      build: function(dialog) {
+        // Find and correct (make work!) any method:delete links
+        dialog.$node.find("[data-method=delete]").on("click", function(e) {
+          e.preventDefault();
+          utilities.post(this.href, { _method: "delete" });
+        });
+      }
+    });
+  }
+}
+
+module.exports = FlowItemMenu
