+++ app/javascript/src/controller_services.js
@@ -25,6 +25,9 @@ const ActivatedMenu = require('./component_activated_menu');
 const DialogApiRequest = require('./component_dialog_api_request');
 const DefaultController = require('./controller_default');
 const ConnectorPath = require('./component_flow_connector_path');
+const FlowItemMenu = require('./component_flow_item_menu');
+const PageAdditionMenu = require('./component_page_addition_menu');
+const ConnectionMenu = require('./component_connection_menu');
 
 const COLUMN_SPACING = 100;
 
@@ -54,6 +57,7 @@ ServicesController.edit = function() {
   createPageAdditionDialog(view);
   createPageAdditionMenu(view);
   createFlowItemMenus(view);
+  createConnectionMenus(view);
 
   if(view.$flowOverview.length) {
     layoutFormFlowOverview(view);
@@ -94,164 +98,164 @@ class FlowItem {
  * ------------------------
  * Control form step (add/edit/delete/preview...) menus
  **/
-class FlowItemMenu extends ActivatedMenu {
-  constructor($node, config) {
-    super($node, mergeObjects({
-      activator_classname: $node.data("activator-classname"),
-      container_id: $node.data("activated-menu-container-id"),
-      activator_text: $node.data("activator-text")
-    }, config));
+// class FlowItemMenu extends ActivatedMenu {
+//   constructor($node, config) {
+//     super($node, mergeObjects({
+//       activator_classname: $node.data("activator-classname"),
+//       container_id: $node.data("activated-menu-container-id"),
+//       activator_text: $node.data("activator-text")
+//     }, config));
 
-    $node.on("menuselect", (event, ui) => {
-      this.selection(event, ui.item);
-    });
+//     $node.on("menuselect", (event, ui) => {
+//       this.selection(event, ui.item);
+//     });
 
-    this.activator.$node.addClass("FlowItemMenuActivator");
-    this.container.$node.addClass("FlowItemMenu");
-    this.uuid = $node.data("uuid");
-    this.title = $node.data("title");
-  }
+//     this.activator.$node.addClass("FlowItemMenuActivator");
+//     this.container.$node.addClass("FlowItemMenu");
+//     this.uuid = $node.data("uuid");
+//     this.title = $node.data("title");
+//   }
 
-   // Handle item selections on the form step context menu elements.
-  selection(event, item) {
-    var action = item.data("action");
+//    // Handle item selections on the form step context menu elements.
+//   selection(event, item) {
+//     var action = item.data("action");
 
-    event.preventDefault();
-    switch(action) {
-      case "preview":
-           this.previewPage(item);
-           break;
+//     event.preventDefault();
+//     switch(action) {
+//       case "preview":
+//            this.previewPage(item);
+//            break;
 
-      case "add":
-           this.addPage(item);
-           break;
+//       case "add":
+//            this.addPage(item);
+//            break;
 
-      case "destination":
-           this.changeDestination(item);
-           break;
+//       case "destination":
+//            this.changeDestination(item);
+//            break;
 
-      case "delete":
-           this.deleteItem(item);
-           break;
+//       case "delete":
+//            this.deleteItem(item);
+//            break;
 
-      case "delete-api":
-           this.deleteItemApi(item);
-           break;
+//       case "delete-api":
+//            this.deleteItemApi(item);
+//            break;
 
-      default: this.link(item);
-    }
-  }
+//       default: this.link(item);
+//     }
+//   }
 
-  link(element) {
-    var $link = element.find("> a");
-    location.href = $link.attr("href");
-  }
+//   link(element) {
+//     var $link = element.find("> a");
+//     location.href = $link.attr("href");
+//   }
 
-  previewPage(element) {
-    var $link = element.find("> a");
-    window.open($link.attr("href"));
-  }
+//   previewPage(element) {
+//     var $link = element.find("> a");
+//     window.open($link.attr("href"));
+//   }
 
-  // Open the views Page Addition Menu
-  addPage(element) {
-    var menu = this._config.view.pageAdditionMenu;
-    menu.addPageAfter = this.uuid;
-    menu.open({
-      my: "left top",
-      at: "right top",
-      of: element
-    });
-  }
+//   // Open the views Page Addition Menu
+//   addPage(element) {
+//     var menu = this._config.view.pageAdditionMenu;
+//     menu.addPageAfter = this.uuid;
+//     menu.open({
+//       my: "left top",
+//       at: "right top",
+//       of: element
+//     });
+//   }
 
-  // Open an API request dialog to change destination
-  changeDestination(element) {
-    var view = this._config.view;
-    var $link = element.find("> a");
-    new DialogApiRequest($link.attr("href"), {
-      activator: $link,
-      buttons: [{
-        text: view.text.dialogs.button_change_destination,
-        click: function(dialog) {
-          dialog.$node.find("form").submit();
-        }
-      }, {
-        text: view.text.dialogs.button_cancel
-      }]
-    });
-  }
+//   // Open an API request dialog to change destination
+//   changeDestination(element) {
+//     var view = this._config.view;
+//     var $link = element.find("> a");
+//     new DialogApiRequest($link.attr("href"), {
+//       activator: $link,
+//       buttons: [{
+//         text: view.text.dialogs.button_change_destination,
+//         click: function(dialog) {
+//           dialog.$node.find("form").submit();
+//         }
+//       }, {
+//         text: view.text.dialogs.button_cancel
+//       }]
+//     });
+//   }
 
-  // Use standard delete modal to remove
-  deleteItem(element) {
-    var view = this._config.view;
-    var $link = element.find("> a");
-    view.dialogConfirmationDelete.open({
-      heading: view.text.dialogs.heading_delete.replace(/%{label}/, this.title),
-      ok: view.text.dialogs.button_delete_page
-      }, function() {
-        post($link.attr("href"), { _method: "delete" });
-    });
-  }
+//   // Use standard delete modal to remove
+//   deleteItem(element) {
+//     var view = this._config.view;
+//     var $link = element.find("> a");
+//     view.dialogConfirmationDelete.open({
+//       heading: view.text.dialogs.heading_delete.replace(/%{label}/, this.title),
+//       ok: view.text.dialogs.button_delete_page
+//       }, function() {
+//         post($link.attr("href"), { _method: "delete" });
+//     });
+//   }
 
-  deleteItemApi(element) {
-    var $link = element.find("> a");
-    new DialogApiRequest($link.attr("href"), {
-      activator: $link,
-      closeOnClickSelector: ".govuk-button",
-      build: function(dialog) {
-        // Find and correct (make work!) any method:delete links
-        dialog.$node.find("[data-method=delete]").on("click", function(e) {
-          e.preventDefault();
-          utilities.post(this.href, { _method: "delete" });
-        });
-      }
-    });
-  }
-}
+//   deleteItemApi(element) {
+//     var $link = element.find("> a");
+//     new DialogApiRequest($link.attr("href"), {
+//       activator: $link,
+//       closeOnClickSelector: ".govuk-button",
+//       build: function(dialog) {
+//         // Find and correct (make work!) any method:delete links
+//         dialog.$node.find("[data-method=delete]").on("click", function(e) {
+//           e.preventDefault();
+//           utilities.post(this.href, { _method: "delete" });
+//         });
+//       }
+//     });
+//   }
+// }
 
 
 /* VIEW SPECIFIC COMPONENT:
  * ------------------------
  * Controls form step Add page functionality
  **/
-class PageAdditionMenu extends ActivatedMenu {
-  constructor($node, config) {
-    super($node, mergeObjects({
-      activator_classname: $node.data("activator-classname"),
-      container_id: $node.data("activated-menu-container-id"),
-      activator_text: $node.data("activator-text")
-    }, config));
+// class PageAdditionMenu extends ActivatedMenu {
+//   constructor($node, config) {
+//     super($node, mergeObjects({
+//       activator_classname: $node.data("activator-classname"),
+//       container_id: $node.data("activated-menu-container-id"),
+//       activator_text: $node.data("activator-text")
+//     }, config));
 
-    this.container.$node.addClass("PageAdditionMenu");
+//     this.container.$node.addClass("PageAdditionMenu");
 
-    // Register event handler for selection of menu item.
-    $node.on("menuselect", (event, ui) => {
-      this.selection(event, ui.item);
-    });
-  }
+//     // Register event handler for selection of menu item.
+//     $node.on("menuselect", (event, ui) => {
+//       this.selection(event, ui.item);
+//     });
+//   }
 
-  set addPageAfter(uuid) {
-    this._uuid = uuid;
-  }
+//   set addPageAfter(uuid) {
+//     this._uuid = uuid;
+//   }
 
-  get addPageAfter() {
-    return this._uuid;
-  }
+//   get addPageAfter() {
+//     return this._uuid;
+//   }
 
-  selection(event, item) {
-    var dialog = this._config.view.pageAdditionDialog;
-    var $form = dialog.$form;
+//   selection(event, item) {
+//     var dialog = this._config.view.pageAdditionDialog;
+//     var $form = dialog.$form;
 
-    // Set the 'add_page_here' value to mark point of new page inclusion.
-    // Should be a uuid of previous page or blank if at end of form.
-    utilities.updateHiddenInputOnForm($form, "page[add_page_after]", this.addPageAfter);
+//     // Set the 'add_page_here' value to mark point of new page inclusion.
+//     // Should be a uuid of previous page or blank if at end of form.
+//     utilities.updateHiddenInputOnForm($form, "page[add_page_after]", this.addPageAfter);
 
-    // Then add any required values.
-    utilities.updateHiddenInputOnForm($form, "page[page_type]", item.data("page-type"));
-    utilities.updateHiddenInputOnForm($form, "page[component_type]", item.data("component-type"));
+//     // Then add any required values.
+//     utilities.updateHiddenInputOnForm($form, "page[page_type]", item.data("page-type"));
+//     utilities.updateHiddenInputOnForm($form, "page[component_type]", item.data("component-type"));
 
-    this._config.view.pageAdditionDialog.open();
-  }
-}
+//     this._config.view.pageAdditionDialog.open();
+//   }
+// }
 
 
 /* VIEW SETUP FUNCTION:
@@ -314,7 +318,19 @@ function createFlowItemMenus(view) {
   });
 }
 
+function createConnectionMenus(view) {
+  $("[data-component='ConnectionMenu']").each((i, el) => {
+    var menu = new ConnectionMenu($(el), {
+      view: view,
+      preventDefault: true, // Stops the default action of triggering element.
+      menu: {
+        position: { at: "right+2 top-2" }
+      }
+    });
 
+    view.addLastPointHandler(menu.activator.$node);
+  });
+}
 /* VIEW SETUP FUNCTION:
  * --------------------
  * Create the main overview layout for form to get the required design.
