+++ app/javascript/styles/_component_activated_menu.scss
@@ -33,6 +33,7 @@
   }
 }
 
+
 .ActivatedMenu_Menu {
   a, span {
     @include button_type_link;
