+++ app/views/services/_flow_layout.html.erb
@@ -63,6 +63,7 @@
           <%= flow_thumbnail_link(item) %>
           <%= flow_text_link(item) %>
           <%= render partial: 'services/flow_page_menu', locals: { item: item } %>
+          <%= render partial: 'services/connection_menu', locals: { item: item } %>
         </article>
       <% end %>
     <% end %>
